Changed getDecreasingSortedDistances function, now it adds distance data to result only if pair index doesn't contain in current Cross-validation group

+++ model.h
@@ -47,7 +47,7 @@ private:
     double window_width;
     bool is_learning;
     DistancesVector distances;
-    QVector<int> accuracy_results;
+    QVector<double> accuracy_results;
     int max_number_of_neighbours = 30;
 
 };
